+++ src/Controller/Mapper/Crud.php
@@ -1,0 +1,22 @@
+<?php
+/**
+ * Bluz Framework Component
+ *
+ * @copyright Bluz PHP Team
+ * @link https://github.com/bluzphp/framework
+ */
+
+/**
+ * @namespace
+ */
+namespace Bluz\Controller\Mapper;
+
+/**
+ * Crud
+ *
+ * @package  Bluz\Rest
+ * @author   Anton Shevchuk
+ */
+class Crud extends AbstractMapper
+{
+}
